requirements.txt
fastapi
pydantic
sqlalchemy
uvicorn
passlib[bcrypt]
python-jose[cryptography]
python-multipart
sqlmodel